requirements.txt
playwright
python-dotenv
streamlit
pyinstaller
